% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcl_utils.r
\name{asin_sqrt_transform}
\alias{asin_sqrt_transform}
\title{Perform Arcsine Square Root Transformation}
\usage{
asin_sqrt_transform(x)

asin_sqrt_transform(x)
}
\arguments{
\item{x}{Numeric vector or matrix}
}
\value{
Transformed data

Transformed data
}
\description{
Perform Arcsine Square Root Transformation

Perform Arcsine Square Root Transformation
}
\keyword{internal}
