% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps2heatmap.R
\name{ps2heatmap}
\alias{ps2heatmap}
\title{Combined p-value and test statistics score "heatmap"}
\usage{
ps2heatmap(R2, P, fontsize = 5, FDR = T, alpha = NA, beta = NA)
}
\arguments{
\item{R2}{A matrix of R2 values}

\item{P}{A matrix of P-values}

\item{fontsize}{The desired font size of the % R2 values in the heatmap.}

\item{FDR}{If \code{T}, P-values are first BH-adjusted, and significance
stars are shown as *** < 0.001, ** < 0.01, * < 0.05.}
}
\value{
ggplot object.
}
\description{
Combined p-value and test statistics score "heatmap"
}
